+++ spec/browser/middleware/basic-auth.spec.js
@@ -16,22 +16,22 @@ describe('Middleware / BasicAuth', () => {
     expect(BasicAuthMiddleware(authData).name).toEqual('BasicAuthMiddleware')
   })
 
-  it('configures the auth data', () => {
-    const newRequest = middleware.request(request)
+  it('configures the auth data', async () => {
+    const newRequest = await middleware.prepareRequest(() => Promise.resolve(request))
     expect(newRequest.auth()).toEqual(authData)
   })
 
-  it('changing the request params does not mutate the configuration', () => {
-    const newRequest = middleware.request(request)
+  it('changing the request params does not mutate the configuration', async () => {
+    const newRequest = await middleware.prepareRequest(() => Promise.resolve(request))
     newRequest.requestParams.auth.password = 'foo'
     expect(authData.password).toEqual('bob')
   })
 
   describe('when the auth property is explicitly defined', () => {
-    it('keeps the original auth data', () => {
+    it('keeps the original auth data', async () => {
       const authData2 = { username: 'bob', password: 'bill' }
       request = new Request(methodDescriptor, { auth: authData2 })
-      const newRequest = middleware.request(request)
+      const newRequest = await middleware.prepareRequest(() => Promise.resolve(request))
       expect(newRequest.auth()).toEqual(authData2)
     })
   })