+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 2.27.0
+
+  - Add `prepareRequest` phase to middleware #129
+
+## 2.26.1
+
+  - Bugfix: Extra query string when path already contains query string #128
+
 ## 2.26.0
 
   - Add async middleware requests support to mocked clients #122